requirements.txt
numpy
Pillow
h5py
tensorflow==2.0.0
tensorflow-hub>=0.2.0